Update license location
Although the license is in the root in the Euphonic repo, in
the tarball distributed on PyPI (which conda-forge pulls from)
it is in the euphonic directory so that it is installed along with
the package. In previous distributions the license was mistakenly
included twice, in the root and in the euphonic directory. It is
now only in the euphonic directory as creation and upload of the
tarball has changed from being done manually to being done by a
Github action.