Remove hard-coded numpy requirement
The hard-coded runtime requirement overrides the conda-forge pinning and results in
euphonic building with the latest version of numpy as a requirement. Conda-forge pins
numpy back further for maximum compatibility so we want to use this pinning to avoid
issues with installing against other packages built on older yet compatible versions of
numpy.